+++ cw-orch-daemon/src/senders/tx.rs
@@ -1,11 +1,14 @@
-use cosmrs::{tx::Msg, AccountId, Any};
+use cosmrs::{
+    tx::{Msg, Raw},
+    AccountId, Any,
+};
 use cosmwasm_std::Addr;
 
-use crate::CosmTxResponse;
+use crate::{cosmos_modules, CosmTxResponse, DaemonError};
 
 use super::query::QuerySender;
 
-pub trait TxSender: QuerySender {
+pub trait TxSender: QuerySender + Sync {
     /// Returns the `AccountId` of the sender that commits the transaction.
     fn account_id(&self) -> AccountId;
 
@@ -34,5 +37,26 @@ pub trait TxSender: QuerySender {
             .unwrap();
 
         self.commit_tx_any(msgs, memo)
+    }
+
+    /// Transaction broadcasting for Tendermint Transactions
+    fn broadcast_tx(
+        &self,
+        tx: Raw,
+    ) -> impl std::future::Future<
+        Output = Result<cosmrs::proto::cosmos::base::abci::v1beta1::TxResponse, DaemonError>,
+    > + Send {
+        async move {
+            let mut client = cosmos_modules::tx::service_client::ServiceClient::new(self.channel());
+            let commit = client
+                .broadcast_tx(cosmos_modules::tx::BroadcastTxRequest {
+                    tx_bytes: tx.to_bytes()?,
+                    mode: cosmos_modules::tx::BroadcastMode::Sync.into(),
+                })
+                .await?;
+
+            let commit = commit.into_inner().tx_response.unwrap();
+            Ok(commit)
+        }
     }
 }
